Remove Limiter interface in favor of more specific RateLimiter.
As it stands, this library only does rate limiting so `Limiter` seems like
a premature generalization. It also forces us to do somewhat magical things
like interface upgrades in order to have a generic interface and still expose
the information we need to set rate limiting heaaders.

This change replaces `Limiter` with a more specific `RateLimiter`. In
the process, it enables limiting a quantity of thing (e.g. megabytes
of data) since that makes particular sense in the context of rate
limiting. In the process, it starts down the path of "emit structs,
consume interfaces" by having `GCRARateLimiter` by a public member
returned by `NewGCRARateLimiter`. This means that if we decide to have
a more generic `Limiter` later, we can simply make the
`GCRARateLimiter` implement it.

It feels like a good sign that this change allowed me to move all of
the tests into a separate `throttled_test` package that only accesses
the public interface.