Redesign `RateQuota` interface to be more intuitive.
With the previous `RateQuota` design, it wasn't intuitive that
`PerMin(30)` would actually allow 59 requests in the first minute--a
burst of 30 plus a steady rate every two seconds. The new parameters,
`MaxRate` and `MaxBurst` separate these into two more intuitive parts.

The new `Rate` type is really just a wrapper for a `time.Duration`.
It is a separate type for two reasons:
* I don't think it's intuitive to describe rates as the time between requests
  at the rate. I think of rates in terms of the inverse of the period.
* Wrapping it in a struct lets us continue support for the deprecated interface.
I regret that this doesn't respect the intelligence of users of the library
but I think it's a reasonable tradeoff. In the next major version, we can
change it to be a `time.Duration`.

See #9 for more discussion beginning at:
https://github.com/throttled/throttled/pull/9#issuecomment-132450629