Fix GCRARateLimiter to properly handle large requests.
The standard GCRA implementation doesn't do a good job handling requests
with a quantity greater than one. Because the limiting is based on the value
before the update, you can make an infinitely large request once and cause
the limiter to essentially lock that key forever.

This modifies the implementation to sanely handle requests with a quantity
greater than one. It doesn't change the behavior at all for requests of
quantity <= 1.