Move store implementations into subpackages of throttled/store.
If store implementations live in the store package, using one of the
implementations means downloading the dependencies for all of them.
By moving to subpackages, you don't need to e.g. have redigo around
to use the MemStore. I used this opportunity to have the `New` methods
for each store return an error.

This would break compatibility guarantees so I've included deprecated shims
in the store package itself.