Rename store subpackages to have a store suffix.
Per comments on #9, `redis` and `mem` are sufficiently short as to be
ambiguous and `redis` even conflicts with the recommended redis library.
Both names are still short with a `store` suffix and there's precendent
for this in the stdlib.

I chose to rename `RedisStore` to `RedigoStore` in the process to avoid
the double-"s" in the `redisstore` subpackage name and clarify that the
implementation is specific to redigo.