Never assign a Redis key's TTL to zero
It seems that under the current implementation we could run into a case
where the TTL is so short that we would never expire certain Redis keys
or might throw an error in the case of `SETEX`.

I think this is mostly relevant for extremely fast rates that would
replenish in less than second, so in practice probably not an issue that
most people would run into.

Here we just broaden the minimum horizon for an expiry out to one
second. Things should still work as expected because the proper TAT
value will still be in there if the key's checked after it should have
theoretically expired, and will also eventually be cleaned up by Redis.

Fixes #26.