Stop using gopkg.in
We originally used `gopkg.in` for the project because V2 introduced
significant breaking changes and providing a way to easily get the V1
API would allow users to continue using that without too much additional
work.

Since then, Go's dependency management systems have advanced somewhat,
especially with the imminent introduction of the new `dep` tool which
allows dependencies to be locked to Git tags. Because `gopkg.in` does
introduce some additional complexity around building and contributing to
the project, this patch drops it in favor of using Git tags combined
with a more modern dependency management tool like `dep`. The V1 API has
also been obsolete for two years now, so it's not likely too many people
are going to need access to it anyway.

My plan here is to get this pull request merged, then tag `master` as
`v2.1.0`.

Fixes #31.