Fixed a bug which occurrs when RemoteAddr is used
The HTTP server in the https://golang.org/pkg/net/http/ package sets RemoteAddr to an "IP:port" address before invoking a handler. The ":port" portion of the RemoteAddr prevents the throtteling from working as expected as soon as VaryBy{} contains "RemoteAddr: true". That's why we need to get rid of the port.