Fix TTL in SetIfNotExistsWithTTL
A time.Duration represents the time as an int64 nanosecond count.
By creating a time.Duration from the amount of seconds in the ttl
we effectively create a time.Duration 1,000,000,000 times smaller
than the orginal ttl, as we're then interpreting the seconds as
nanoseconds.

The following playground example illustrates this:
https://play.golang.org/p/8w15MOkAfC3