Add context objects to goredis calls
The build is currently broken because goredis changed their API in a few
ways -- mostly adding context objects to all calls, but also with a few
minor type changes.

We'll want to have a better versioning strategy at some point here, but
for now, we update our goredis adapter to be compatible.