Revert upgrade to go-redis v8
Reverts the upgrade to go-redis v8 that I made a few days ago because in
retrospect, it seems likely to cause more problems than it solves.

I'd thought that going to v8 was somewhat unpreventable because it was
their `master` branch, but the main problem seems to have been that this
project was not on Go Modules. Now that it is, go-redis v6 gets pulled
by default, and an explicit major needs to be specified to get anything
later.

Also, according to the go-redis CHANGELOG, v8 is not yet released.

I'm not sure what the best course of action is going forward, but it may
be that we should release new stores to support new majors from the
supported adapters. There may be a better way, but I'm not sure what it
is yet.