Have Throttled imported via a `/v2` in path
Playing around with some demo projects, I found that Go Modules was
still not playing nicely with Throttled because it expected the major
version in the path to match the major version in tags, and Throttled
was already on V2 long before Go Modules were a thing.

This is unfortunate since it's going to cause pain to existing users who
try to upgrade, but I think the way to go is probably to recommend the
use of imports with `/v2` in the path. I've updated all imports,
comments, and the README to reflect this.