Drop Go versions 1.13, 1.14, 1.15, 1.16, and 1.17 from build matrix
Proximate impetus is trying to support go-redis.v9 in #103, which uses
some more modern Go conventions like `strings.Cut` and `time.UnixMillis`
which won't build prior to 1.18 and 1.17 respectively.

Go technically supports only the last two versions (1.19 and 1.20), and
1.21 has a release candidate so even 1.19 is about to drop. Let's try
not to be quite that aggressive, but drop everything up to 1.18.